+++ scripts/var_mk_yocto_sdcard/variscite_scripts/mx6ul_mx7_install_yocto.sh
@@ -244,10 +244,14 @@ install_rootfs_to_emmc()
 	done
 
 	# Adjust u-boot-fw-utils for eMMC on the installed rootfs
-	rm ${mountdir_prefix}${rootfspart}/sbin/fw_printenv-nand
-	mv ${mountdir_prefix}${rootfspart}/sbin/fw_printenv-mmc ${mountdir_prefix}${rootfspart}/sbin/fw_printenv
-	sed -i "/mtd/ s/^#*/#/" ${mountdir_prefix}${rootfspart}/etc/fw_env.config
-	sed -i "s/#*\/dev\/mmcblk./\/dev\/${block}/" ${mountdir_prefix}${rootfspart}/etc/fw_env.config
+	if [[ -e ${mountdir_prefix}${rootfspart}/sbin/fw_printenv-nand && \
+	      -e ${mountdir_prefix}${rootfspart}/sbin/fw_printenv-mmc && \
+	      -e ${mountdir_prefix}${rootfspart}/etc/fw_env.config ]] ; then
+		rm ${mountdir_prefix}${rootfspart}/sbin/fw_printenv-nand
+		mv ${mountdir_prefix}${rootfspart}/sbin/fw_printenv-mmc ${mountdir_prefix}${rootfspart}/sbin/fw_printenv
+		sed -i "/mtd/ s/^#*/#/" ${mountdir_prefix}${rootfspart}/etc/fw_env.config
+		sed -i "s/#*\/dev\/mmcblk./\/dev\/${block}/" ${mountdir_prefix}${rootfspart}/etc/fw_env.config
+	fi
 
 	echo
 	sync
